+++ packages/core-js/full/symbol/index.js
@@ -1,5 +1,7 @@
 var parent = require('../../actual/symbol');
 require('../../modules/esnext.symbol.async-dispose');
+require('../../modules/esnext.symbol.is-registered');
+require('../../modules/esnext.symbol.is-well-known');
 require('../../modules/esnext.symbol.matcher');
 require('../../modules/esnext.symbol.metadata-key');
 require('../../modules/esnext.symbol.observable');