+++ tests/entries/unit.mjs
@@ -814,6 +814,8 @@ for (PATH of ['core-js-pure', 'core-js']) {
     ` === 'a1b');
     ok('next' in load(NS, 'string/code-points')('a'));
     ok('next' in load(NS, 'string/virtual/code-points').call('a'));
+    ok(load(NS, 'symbol/is-registered')(1) === false);
+    ok(load(NS, 'symbol/is-well-known')(1) === false);
     ok(load(NS, 'symbol/matcher'));
     ok(load(NS, 'symbol/metadata'));
     ok(load(NS, 'symbol/metadata-key'));
@@ -940,6 +942,7 @@ for (PATH of ['core-js-pure', 'core-js']) {
   load('proposals/string-replace-all');
   load('proposals/string-replace-all-stage-4');
   load('proposals/symbol-description');
+  load('proposals/symbol-predicates');
   load('proposals/url');
   load('proposals/using-statement');
   load('proposals/well-formed-stringify');